+++ src/app/Fake.Core.Process/RawProc.fs
@@ -39,7 +39,9 @@ module EnvMap =
             IMap.empty
 
     let ofSeq l : EnvMap =
-        empty.AddRange(l |> Seq.map (fun (k, v) -> KeyValuePair<_, _>(k, v)))
+        empty
+            .AddRange(Environment.environVars () |> Seq.map (fun (k, v) -> KeyValuePair<_,_>(k, v)))
+            .AddRange(l |> Seq.map (fun (k, v) -> KeyValuePair<_,_>(k, v)))
 
     let create () = ofSeq (Environment.environVars ())